- Updated import paths for `OptionOrder` and `StoredPrice` to reflect new directory structure, enhancing clarity and maintainability. - Deleted obsolete `misc.ts` and `order.ts` type definitions, streamlining the type definitions in the project. - Consolidated `StoredPrice` type definition into `priceMarketTypes.ts` for better organization of market-related types.
